Refactoring: Move messages generation into separate file.
